+++ inst/SQL/drug_prescriptions_edges_SQLite.sql
@@ -80,7 +80,7 @@ FROM (
 
 WHERE (final.edge_type = 'combination')
    OR (final.edge_type = 'continuation'
-    AND (final.[from_status] NOT IN ('cancelled', 'draft', 'entered-in-error')
-        AND final.[to_status] NOT IN ('cancelled', 'draft', 'entered-in-error')))
+    AND (final.[from_status] NOT IN ('unknown', 'cancelled', 'draft', 'entered-in-error')
+        AND final.[to_status] NOT IN ('unknown', 'cancelled', 'draft', 'entered-in-error')))
 
 ORDER BY [patient_id], [from_id], [from_authoring], [to_start];